% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_cache.R
\name{syncGdCache}
\alias{syncGdCache}
\title{Title}
\usage{
syncGdCache(local_folder, gdrive_folder)
}
\arguments{
\item{local_folder}{The full path to local cloud cache}

\item{gdrive_folder}{The google drive folder id like "https://drive.google.com/drive/u/3/folders/<here is folder id>"}
}
\value{

}
\description{
Title
}
\examples{
syncGdCache(here::here("cloud_data"), gdrive_folder = "1pMs9-auXmGsw9OGZD-9KiCCEaPryZ52s")
}
